add ocp config test and improve makefile
